fix(opus): the container whose channels name no order, and the pre-roll asking early with nothing to trim
Three the previous commit's own review turned up, none of them visible to a
green suite.

Opus in Matroska stopped decoding. `plane_sources` took the container's
`Channels` and required a positioned one, where the range test it replaced
accepted any set of count one or two. `symphonia-format-mkv` reports
`Channels::Discrete` for every audio track it demuxes, so a stereo
Opus-in-`.mka` that played before the surround work failed to build a decoder
and was silently skipped, with `.mka` in `AUDIO_EXTENSIONS`, Matroska named in
the README and in this module's own doc, and `ffmpeg -c:a libopus -f matroska`
producing such a file in one command. Nothing could see it: the fixture walk
opens `silence.mka`, which carries FLAC.

The set is ours to state, not only to check, for the mirror of the reason the
stated stream layout already was. `positioned_layout` derives both the
positioned set and the plane order from `VORBIS_ORDER` against the count, which
answers for all three containers and deletes the comparison rather than adding
a branch beside it: Ogg and MP4 build their sets from that same table, so it
could only ever have fired for the set that was breaking. It drops a `Channels`
clone per decoder build too, that type having a boxed variant. A `.mka` fixture
carrying Opus is the pin still owed.

A pre-roll with no timebase replayed 80 ms rather than warming anything.
`samples_before` and `trim_target` both return early without one, so the seek
asked short and nothing took the difference back off. Unreachable while Ogg,
MP4 and Matroska all state a timebase, and free to close by answering the
timebase before the pre-roll.

`parse_r128_gain` parsed `i32`, so a tag of "2000000" reached the playback DSP
as roughly 7800 dB. RFC 7845 section 5.2 states the field as `i16`, which caps
the answer at 128 dB either way and is the type to parse.

One comment was wrong rather than merely long. Symphonia's
`copy_from_slice_interleaved` copies planes in pairs, so the de-interleave does
not walk the same samples in the same order for the counts where the two orders
agree; it gives that pass up, which is what the note says now. The loop asks
the iterator for the stride instead of indexing the frame, which drops a bounds
check from the inner loop. The rest of the file's prose is the same arguments in
fewer lines.

+++ crates/melodia-store/src/media/ingest/metadata.rs
@@ -178,9 +178,10 @@ const R128_TO_REPLAYGAIN_DB: f64 = 5.0;
 /// Parse an `R128_*_GAIN` value — Q7.8 fixed-point dB, e.g. "-1280" — restated against
 /// `ReplayGain`'s reference so it can share the columns and the DSP.
 ///
-/// No non-finite guard, unlike the two above: an integer scaled and offset cannot be one.
+/// `i16` because RFC 7845 §5.2 states the field as one, which caps the answer at ±128 dB: no
+/// non-finite guard is owed, unlike the two above, and no absurd one can reach the DSP either.
 fn parse_r128_gain(s: &str) -> Option<f64> {
-    let q7_8 = s.trim().parse::<i32>().ok()?;
+    let q7_8 = s.trim().parse::<i16>().ok()?;
     Some(f64::from(q7_8) / 256.0 + R128_TO_REPLAYGAIN_DB)
 }
